fix(ci): resolve cargo fmt + clippy lint failures
- Run cargo fmt --all to normalize formatting
- Fix 50 clippy lints (needless_borrow, collapsible_if, io_other_error,
  is_multiple_of, redundant_pattern_matching, useless vec)
- These were breaking the lint CI job

+++ src/resp.rs
@@ -103,7 +103,7 @@ pub fn parse_command(buf: &[u8]) -> Result<Option<(Vec<String>, usize)>, RespErr
             Some(v) => v,
             None => return Ok(None),
         };
-        let bulk_len = parse_int(len_line)? as isize;
+        let bulk_len = parse_int(len_line)?;
         if bulk_len < 0 {
             parts.push(String::new());
             offset = offset + 1 + len_consumed;
